Add Elementary Parser, ErrorSys, and Improve Scanner (#1)
This has multiple benefits, most notably is the easier tracking of line
numbers for the error system. Another benefit of this is that it'll be
more performant to grab line-by-line batches, as before either way we'd
be stepping char-by char which is unneccessary and would cause much more
logic in determining line count. All around, this is a good commit to
clean up Scanner and fix the line bug.

Another change that I snuck in here is that each Token now carries it's
line number, so when a token causes an error it hold's the line number
of it's original location. This prevents us from grabbing incorrect line
numbers from grabbing it in a different scanner state.