Add validation stage for futher semantic analysis
After the first parse, the validation stage will now run further checks
to ensure that the source language makes sense before continuing to code
generation. In this change multiple other issues had to be addressed.
The last commit involved function return types no longer being assumed,
but native decls were overlooked in that change and is included in this
one.

Logic now exists to ensure that functions must be declared native or
contain a definition before their use, and tests have been added as
always. Older tests had to be updated due to this as well.