Fix windows script compilation bug
Other compilers (gcc & clang) will zero-initialze class members,
probably out of their own understanding that *not* doing it is a
mistake. Having relied on this, it seems that that bad behavior finally
caught up to me. I was awarded this bug for our windows builds that I
spent hours to track down. The more you know!